#191 Extract common logic to helper file
