+++ Server/ConcordiaCurriculumManagerTest/UnitTests/Controller/CourseControllerTest.cs
@@ -1,8 +1,6 @@
 ﻿using AutoMapper;
 using ConcordiaCurriculumManager.Controllers;
 using ConcordiaCurriculumManager.Models.Curriculum.Dossiers;
-using ConcordiaCurriculumManager.Models.Curriculum;
-using ConcordiaCurriculumManager.Models.Users;
 using ConcordiaCurriculumManager.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -10,8 +8,8 @@ using Moq;
 using System.Net;
 using ConcordiaCurriculumManager.DTO.Courses;
 using ConcordiaCurriculumManager.Filters.Exceptions;
-using ConcordiaCurriculumManager.DTO.Dossiers.CourseRequests.InputDTOs;
 using ConcordiaCurriculumManager.DTO.Dossiers.CourseRequests.OutputDTOs;
+using ConcordiaCurriculumManagerTest.UnitTests.UtilityFunctions;
 
 namespace ConcordiaCurriculumManagerTest.UnitTests.Controller;
 
@@ -38,12 +36,12 @@ public class CourseControllerTest
     [TestMethod]
     public async Task InitiatiateCourseCreation_ValidCall_ReturnsData()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseCreationInitiationDTO = GetSampleCourseCreationInitiationDTO(course, dossier);
-        var courseCreationRequestDTO = GetSampleCourseCreationRequestDTO(course, dossier);
-        var courseCreationRequest = GetSampleCourseCreationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseCreationInitiationDTO = TestData.GetSampleCourseCreationInitiationDTO(dossier);
+        var courseCreationRequestDTO = TestData.GetSampleCourseCreationRequestDTO(course, dossier);
+        var courseCreationRequest = TestData.GetSampleCourseCreationRequest(dossier, course);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseCreation(courseCreationInitiationDTO, user.Id)).ReturnsAsync(courseCreationRequest);
@@ -60,12 +58,12 @@ public class CourseControllerTest
     [ExpectedException(typeof(BadRequestException))]
     public async Task InitiatiateCourseCreation_InvalidCall_ThrowsException()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseCreationInitiationDTO = GetSampleCourseCreationInitiationDTO(course, dossier);
-        var courseCreationRequestDTO = GetSampleCourseCreationRequestDTO(course, dossier);
-        var courseCreationRequest = GetSampleCourseCreationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseCreationInitiationDTO = TestData.GetSampleCourseCreationInitiationDTO(dossier);
+        var courseCreationRequestDTO = TestData.GetSampleCourseCreationRequestDTO(course, dossier);
+        var courseCreationRequest = TestData.GetSampleCourseCreationRequest(dossier, course);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseCreation(courseCreationInitiationDTO, user.Id)).Throws(new BadRequestException());
@@ -77,12 +75,12 @@ public class CourseControllerTest
     [TestMethod]
     public async Task InitiatiateCourseModification_ValidCall_ReturnsData()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseModificationInitiationDTO = GetSampleCourseModificationInitiationDTO(course, dossier);
-        var courseModificationRequestDTO = GetSampleCourseModificationRequestDTO(course, dossier);
-        var courseModificationRequest = GetSampleCourseModificationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseModificationInitiationDTO = TestData.GetSampleCourseCreationModificationDTO(course, dossier);
+        var courseModificationRequestDTO = TestData.GetSampleCourseModificationRequestDTO(course, dossier);
+        var courseModificationRequest = TestData.GetSampleCourseModificationRequest(dossier, course);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseModification(courseModificationInitiationDTO, user.Id)).ReturnsAsync(courseModificationRequest);
@@ -99,12 +97,12 @@ public class CourseControllerTest
     [ExpectedException(typeof(NotFoundException))]
     public async Task InitiatiateCourseModification_InvalidCall_ThrowsException()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseModificationInitiationDTO = GetSampleCourseModificationInitiationDTO(course, dossier);
-        var courseModificationRequestDTO = GetSampleCourseModificationRequestDTO(course, dossier);
-        var courseModificationRequest = GetSampleCourseModificationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseModificationInitiationDTO = TestData.GetSampleCourseCreationModificationDTO(course, dossier);
+        var courseModificationRequestDTO = TestData.GetSampleCourseModificationRequestDTO(course, dossier);
+        var courseModificationRequest = TestData.GetSampleCourseModificationRequest(dossier, course);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseModification(courseModificationInitiationDTO, user.Id)).Throws(new NotFoundException());
@@ -116,12 +114,12 @@ public class CourseControllerTest
     [TestMethod]
     public async Task InitiatiateCourseDeletion_ValidCall_ReturnsData()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseDeletionInitiationDTO = GetSampleCourseDeletionInitiationDTO(course, dossier);
-        var courseDeletionnRequestDTO = GetSampleCourseDeletionRequestDTO(course, dossier);
-        var courseDeletionRequest = GetSampleCourseDeletionRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseDeletionInitiationDTO = TestData.GetSampleCourseCreationDeletionDTO(course, dossier);
+        var courseDeletionnRequestDTO = TestData.GetSampleCourseDeletionRequestDTO(course, dossier);
+        var courseDeletionRequest = TestData.GetSampleCourseDeletionRequest(course, dossier);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseDeletion(courseDeletionInitiationDTO, user.Id)).ReturnsAsync(courseDeletionRequest);
@@ -138,12 +136,12 @@ public class CourseControllerTest
     [ExpectedException(typeof(NotFoundException))]
     public async Task InitiatiateCourseDeletion_InvalidCall_ThrowsException()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseDeletionInitiationDTO = GetSampleCourseDeletionInitiationDTO(course, dossier);
-        var courseDeletionRequestDTO = GetSampleCourseDeletionRequestDTO(course, dossier);
-        var courseDeletionRequest = GetSampleCourseDeletionRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseDeletionInitiationDTO = TestData.GetSampleCourseCreationDeletionDTO(course, dossier);
+        var courseDeletionnRequestDTO = TestData.GetSampleCourseDeletionRequestDTO(course, dossier);
+        var courseDeletionRequest = TestData.GetSampleCourseDeletionRequest(course, dossier);
 
         _userAuthenticationService.Setup(x => x.GetCurrentUserClaim(It.IsAny<string>())).Returns(user.Id.ToString);
         _courseService.Setup(x => x.InitiateCourseDeletion(courseDeletionInitiationDTO, user.Id)).Throws(new NotFoundException());
@@ -160,7 +158,7 @@ public class CourseControllerTest
     {
         var subject = "SOEN";
         var catalog = "490";
-        var course = GetSampleCourse();
+        var course = TestData.GetSampleCourse();
        _courseService.Setup(cr => cr.GetCourseData(subject, catalog)).ReturnsAsync(course);
 
         var actionResult = await _courseController.GetCourseData(subject, catalog);
@@ -185,12 +183,12 @@ public class CourseControllerTest
     [TestMethod]
     public async Task EditCourseCreationRequest_ValidCall_ReturnsData()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseCreationRequestDTO = GetSampleCourseCreationRequestDTO(course, dossier);
-        var courseCreationRequest = GetSampleCourseCreationRequest(course, dossier);
-        var editCourseCreationRequestDTO = GetSampleEditCourseCreationRequestDTO();
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseCreationRequestDTO = TestData.GetSampleCourseCreationRequestDTO(course, dossier);
+        var courseCreationRequest = TestData.GetSampleCourseCreationRequest(dossier, course);
+        var editCourseCreationRequestDTO = TestData.GetSampleEditCourseCreationRequestDTO();
 ;
         _courseService.Setup(x => x.EditCourseCreationRequest(editCourseCreationRequestDTO)).ReturnsAsync(courseCreationRequest);
         _mapper.Setup(x => x.Map<CourseCreationRequestDTO>(It.IsAny<CourseCreationRequest>())).Returns(courseCreationRequestDTO);
@@ -206,7 +204,7 @@ public class CourseControllerTest
     [ExpectedException(typeof(NotFoundException))]
     public async Task EditCourseCreationRequest_InvalidCall_ThrowsException()
     {
-        var editCourseCreationRequestDTO = GetSampleEditCourseCreationRequestDTO();
+        var editCourseCreationRequestDTO = TestData.GetSampleEditCourseCreationRequestDTO();
 
         _courseService.Setup(x => x.EditCourseCreationRequest(editCourseCreationRequestDTO)).Throws(new NotFoundException());
 
@@ -217,12 +215,12 @@ public class CourseControllerTest
     [TestMethod]
     public async Task EditCourseModificationRequest_ValidCall_ReturnsData()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseModificationRequestDTO = GetSampleCourseModificationRequestDTO(course, dossier);
-        var courseModificationRequest = GetSampleCourseModificationRequest(course, dossier);
-        var editCourseModificationRequestDTO = GetSampleEditCourseModificationRequestDTO();
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseModificationRequestDTO = TestData.GetSampleCourseModificationRequestDTO(course, dossier);
+        var courseModificationRequest = TestData.GetSampleCourseModificationRequest(dossier, course);
+        var editCourseModificationRequestDTO = TestData.GetSampleEditCourseModificationRequestDTO();
 
         _courseService.Setup(x => x.EditCourseModificationRequest(editCourseModificationRequestDTO)).ReturnsAsync(courseModificationRequest);
         _mapper.Setup(x => x.Map<CourseModificationRequestDTO>(It.IsAny<CourseModificationRequest>())).Returns(courseModificationRequestDTO);
@@ -238,7 +236,7 @@ public class CourseControllerTest
     [ExpectedException(typeof(NotFoundException))]
     public async Task EditCourseModificationRequest_InvalidCall_ThrowsException()
     {
-        var editCourseModificationRequestDTO = GetSampleEditCourseModificationRequestDTO();
+        var editCourseModificationRequestDTO = TestData.GetSampleEditCourseModificationRequestDTO();
 
         _courseService.Setup(x => x.EditCourseModificationRequest(editCourseModificationRequestDTO)).Throws(new NotFoundException());
 
@@ -252,10 +250,10 @@ public class CourseControllerTest
     [TestMethod]
     public async Task DeleteCourseCreationRequest_ValidCall_Returns204()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseCreationRequest = GetSampleCourseCreationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseCreationRequest = TestData.GetSampleCourseCreationRequest(dossier, course);
 
         var actionResult = await _courseController.DeleteCourseCreationRequest(courseCreationRequest.Id);
         var objectResult = (NoContentResult)actionResult;
@@ -275,10 +273,10 @@ public class CourseControllerTest
     [TestMethod]
     public async Task DeleteCourseModificationRequest_ValidCall_Returns204()
     {
-        var user = GetSampleUser();
-        var dossier = GetSampleDossier(user);
-        var course = GetSampleCourse();
-        var courseModificationRequest = GetSampleCourseModificationRequest(course, dossier);
+        var user = TestData.GetSampleUser();
+        var dossier = TestData.GetSampleDossier(user);
+        var course = TestData.GetSampleCourse();
+        var courseModificationRequest = TestData.GetSampleCourseModificationRequest(dossier, course);
 
         var actionResult = await _courseController.DeleteCourseModificationRequest(courseModificationRequest.Id);
         var objectResult = (NoContentResult)actionResult;
@@ -293,226 +291,5 @@ public class CourseControllerTest
         _courseService.Setup(service => service.DeleteCourseModificationRequest(It.IsAny<Guid>())).Throws(new Exception());
 
         await _courseController.DeleteCourseModificationRequest(It.IsAny<Guid>());
-    }
-
-
-    private User GetSampleUser()
-    {
-        return new User
-        {
-            Id = new Guid(),
-            FirstName = "Joe",
-            LastName = "Smith",
-            Email = "jsmith@ccm.com",
-            Password = "Password123!"
-        };
-    }
-
-    private Dossier GetSampleDossier(User user)
-    {
-        return new Dossier
-        {
-            Initiator = user,
-            InitiatorId = user.Id,
-            Title = "Dossier 1",
-            Description = "Text description of a dossier.",
-            Published = false,
-        };
-    }
-
-    private Course GetSampleCourse()
-    {
-        var id = Guid.NewGuid();
-
-        return new Course
-        {
-            Id = id,
-            CourseID = 1000,
-            Subject = "SOEN",
-            Catalog = "490",
-            Title = "Capstone",
-            Description = "Curriculum manager building simulator",
-            CreditValue = "6",
-            PreReqs = "SOEN 390",
-            CourseNotes = "Lots of fun",
-            Career = CourseCareerEnum.UGRD,
-            EquivalentCourses = "",
-            CourseState = CourseStateEnum.NewCourseProposal,
-            Version = 1,
-            Published = true,
-            CourseCourseComponents = CourseCourseComponent.GetComponentCodeMapping(new Dictionary<ComponentCodeEnum, int?> { { ComponentCodeEnum.LEC, 3 } }, id)
-        };
-    }
-
-    private CourseCreationInitiationDTO GetSampleCourseCreationInitiationDTO(Course course, Dossier dossier)
-    {
-        return new CourseCreationInitiationDTO
-        {
-            Subject = "SOEN",
-            Catalog = "490",
-            Title = "Capstone",
-            Description = "Curriculum manager building simulator",
-            CreditValue = "6",
-            PreReqs = "SOEN 390",
-            CourseNotes = "Fun",
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Career = CourseCareerEnum.UGRD,
-            EquivalentCourses = "",
-            ComponentCodes = new Dictionary<ComponentCodeEnum, int?> { { ComponentCodeEnum.LEC, 3 } },
-            SupportingFiles = new Dictionary<string, string> { { "name.pdf", "base64content" } },
-            DossierId = dossier.Id
-        };
-    }
-
-    private CourseCreationRequestDTO GetSampleCourseCreationRequestDTO(Course course, Dossier dossier)
-    {
-        return new CourseCreationRequestDTO
-        {
-            Id = Guid.NewGuid(),
-            DossierId = dossier.Id,
-            NewCourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private CourseCreationRequest GetSampleCourseCreationRequest(Course course, Dossier dossier)
-    {
-        return new CourseCreationRequest
-        {
-            DossierId = dossier.Id,
-            NewCourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private CourseModificationInitiationDTO GetSampleCourseModificationInitiationDTO(Course course, Dossier dossier)
-    {
-        return new CourseModificationInitiationDTO
-        {
-            Subject = "SOEN",
-            Catalog = "490",
-            Title = "Capstone",
-            Description = "Curriculum manager building simulator",
-            CreditValue = "6",
-            PreReqs = "SOEN 390",
-            CourseNotes = "Fun",
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Career = CourseCareerEnum.UGRD,
-            EquivalentCourses = "",
-            ComponentCodes = new Dictionary<ComponentCodeEnum, int?> { { ComponentCodeEnum.LEC, 3 } },
-            SupportingFiles = new Dictionary<string, string> { { "name.pdf", "base64content" } },
-            DossierId = dossier.Id,
-            CourseId = 1,
-        };
-    }
-
-    private CourseModificationRequestDTO GetSampleCourseModificationRequestDTO(Course course, Dossier dossier)
-    {
-        return new CourseModificationRequestDTO
-        {
-            Id = Guid.NewGuid(),
-            DossierId = dossier.Id,
-            CourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private CourseModificationRequest GetSampleCourseModificationRequest(Course course, Dossier dossier)
-    {
-        return new CourseModificationRequest
-        {
-            DossierId = dossier.Id,
-            CourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private CourseDeletionInitiationDTO GetSampleCourseDeletionInitiationDTO(Course course, Dossier dossier)
-    {
-        return new CourseDeletionInitiationDTO
-        {
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            DossierId = dossier.Id,
-            Subject = "SOEN",
-            Catalog = "490"
-        };
-    }
-
-    private CourseDeletionRequestDTO GetSampleCourseDeletionRequestDTO(Course course, Dossier dossier)
-    {
-        return new CourseDeletionRequestDTO
-        {
-            Id = Guid.NewGuid(),
-            DossierId = dossier.Id,
-            CourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private CourseDeletionRequest GetSampleCourseDeletionRequest(Course course, Dossier dossier)
-    {
-        return new CourseDeletionRequest
-        {
-            DossierId = dossier.Id,
-            CourseId = course.Id,
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Comment = "Fun",
-        };
-    }
-
-    private EditCourseCreationRequestDTO GetSampleEditCourseCreationRequestDTO()
-    {
-        return new EditCourseCreationRequestDTO
-        {
-            Id = Guid.NewGuid(),
-            DossierId = Guid.NewGuid(),
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Subject = "SOEN Modified3",
-            Catalog = "500",
-            Title = "Super Capstone for Software Engineers",
-            Description = "An advanced capstone project for final year software engineering students.",
-            CourseNotes = "Students are required to present their project at the end of the semester.",
-            CreditValue = "6.5",
-            PreReqs = "SOEN 490 previously or concurrently",
-            Career = CourseCareerEnum.UGRD,
-            EquivalentCourses = "SOEN 499",
-            ComponentCodes = new Dictionary<ComponentCodeEnum, int?> { { ComponentCodeEnum.LEC, 3 } },
-            SupportingFiles = new Dictionary<string, string> { { "name.pdf", "base64content" } },
-        };
-    }
-
-    private EditCourseModificationRequestDTO GetSampleEditCourseModificationRequestDTO()
-    {
-        return new EditCourseModificationRequestDTO
-        {
-            Id = Guid.NewGuid(),
-            DossierId = Guid.NewGuid(),
-            Rationale = "It's necessary",
-            ResourceImplication = "New prof needed",
-            Title = "Super Capstone for Software Engineers",
-            Description = "An advanced capstone project for final year software engineering students.",
-            CourseNotes = "Students are required to present their project at the end of the semester.",
-            CreditValue = "6.5",
-            PreReqs = "SOEN 490 previously or concurrently",
-            Career = CourseCareerEnum.UGRD,
-            EquivalentCourses = "SOEN 499",
-            ComponentCodes = new Dictionary<ComponentCodeEnum, int?> { { ComponentCodeEnum.LEC, 3 } },
-            SupportingFiles = new Dictionary<string, string> { { "name.pdf", "base64content" } },
-        };
     }
 }
